refactor: move table functions to one file
As mentioned by @benjaminwinger, since these functions are only
reference in this source file, they need not be in headers, and can all
be completely inlined in the source file instead. Ref #2142.
