+++ src/function/CMakeLists.txt
@@ -1,5 +1,4 @@
 add_subdirectory(aggregate)
-add_subdirectory(table_functions)
 
 add_library(kuzu_function
         OBJECT
